+++ src/distro/debian.sh
@@ -45,7 +45,11 @@ generate_isowork(){
         echo "insmod all_video" > isowork/boot/grub/grub.cfg
     fi
     mkdir -p isowork/live/ || true
-    mv filesystem.squashfs isowork/live/
+    if [[ -e "filesystem.squashfs" ]]; then
+        mv filesystem.squashfs isowork/live/
+    elif [[ -e "filesystem.erofs" ]]; then
+        mv filesystem.erofs isowork/live/
+    fi
     ls isowork/boot/ | grep "vmlinuz" | while read line ; do
         echo "menuentry $(distro_name) --class debian {" >> isowork/boot/grub/grub.cfg
         echo "  linux /boot/$line boot=live" >> isowork/boot/grub/grub.cfg
@@ -61,6 +65,6 @@ customize_airootfs(){
 clear_rootfs(){
     run_in_chroot apt clean
     run_in_chroot apt autoremove
-    rm -rf $rootfs/var/lib/apt/lists
-    find "$rootfs/var/log/" -type f | xargs rm -f
+    rm -rf $rootfs/var/lib/apt/lists || true
+    find "$rootfs/var/log/" -type f | xargs rm -f  || true
 }
